Segmenting and Clustering Neighborhoods in the city of Toronto, Canada
IBM Applied Data Science  Capstone Project : For the Toronto neighborhood data, a Wikipedia page exists that has all the information we need to explore and cluster the neighborhoods in Toronto. Scrape the Wikipedia page and wrangle the data, clean it, and then read it into a pandas dataframe so that it is in a structured format.Once the data is in a structured format, it was explored and  the neighborhoods in the city of Toronto was clustered.